refactor: split auth.rs into auth/ module
Extract monolithic auth.rs (1362 lines) into focused submodules:
- auth/mod.rs: middleware, AuthFailureTracker, path helpers
- auth/htpasswd.rs: HtpasswdAuth provider
- auth/token_routes.rs: token CRUD API handlers

Zero behavior change — mechanical split to prepare for OIDC
provider addition. All 74 auth tests pass unchanged.

Refs: #342
