+++ nora-registry/src/auth/mod.rs
@@ -1,6 +1,19 @@
 // Copyright (c) 2026 The NORA Authors
 // SPDX-License-Identifier: MIT
 
+//! Authentication module — middleware, providers, and token routes.
+//!
+//! Supports:
+//! - Basic auth via htpasswd files
+//! - Bearer token auth (opaque tokens with Argon2 verification)
+//! - Brute-force protection with exponential backoff
+
+mod htpasswd;
+mod token_routes;
+
+pub use htpasswd::HtpasswdAuth;
+pub use token_routes::{token_routes, TokenListItem, TokenListResponse};
+
 use axum::{
     body::Body,
     extract::{ConnectInfo, State},
@@ -11,11 +24,9 @@ use axum::{
 use base64::{engine::general_purpose::STANDARD, Engine};
 use std::collections::HashMap;
 use std::net::{IpAddr, SocketAddr};
-use std::path::Path;
 use std::sync::Arc;
 use std::time::Instant;
 
-use crate::tokens::Role;
 use crate::AppState;
 
 /// Tracks failed authentication attempts per IP for brute-force protection.
@@ -78,50 +89,6 @@ impl AuthFailureTracker {
     }
 }
 
-/// Htpasswd-based authentication
-#[derive(Clone)]
-pub struct HtpasswdAuth {
-    users: HashMap<String, String>, // username -> bcrypt hash
-}
-
-impl HtpasswdAuth {
-    /// Load users from htpasswd file
-    pub fn from_file(path: &Path) -> Option<Self> {
-        let content = std::fs::read_to_string(path).ok()?;
-        let mut users = HashMap::new();
-
-        for line in content.lines() {
-            let line = line.trim();
-            if line.is_empty() || line.starts_with('#') {
-                continue;
-            }
-            if let Some((username, hash)) = line.split_once(':') {
-                users.insert(username.to_string(), hash.to_string());
-            }
-        }
-
-        if users.is_empty() {
-            None
-        } else {
-            Some(Self { users })
-        }
-    }
-
-    /// Verify username and password
-    pub fn authenticate(&self, username: &str, password: &str) -> bool {
-        if let Some(hash) = self.users.get(username) {
-            bcrypt::verify(password, hash).unwrap_or(false)
-        } else {
-            false
-        }
-    }
-
-    /// Get list of usernames
-    pub fn list_users(&self) -> Vec<&str> {
-        self.users.keys().map(|s| s.as_str()).collect()
-    }
-}
-
 /// Check if path is public (no auth required)
 fn is_public_path(path: &str) -> bool {
     // Token UI pages require auth — exclude before wildcard match
@@ -154,10 +121,7 @@ fn is_docker_auth_challenge_path(path: &str) -> bool {
 /// If the direct peer IP is not in `trusted_proxies`, XFF/X-Real-IP headers are
 /// ignored and the peer IP is returned. This prevents attackers from spoofing
 /// their IP to bypass `AuthFailureTracker` lockout.
-/// Resolve the real client IP from peer address and forwarding headers.
-/// If the peer is a trusted proxy, checks X-Forwarded-For and X-Real-IP.
-/// Otherwise returns the peer IP directly (prevents spoofing).
-fn resolve_client_ip(
+pub(crate) fn resolve_client_ip(
     peer: IpAddr,
     headers: &HeaderMap,
     trusted_proxies: &crate::config::TrustedProxies,
@@ -356,312 +320,10 @@ fn unauthorized_response(message: &str, realm: &str) -> Response {
         .into_response()
 }
 
-// Token management API routes
-use axum::{routing::post, Json, Router};
-use serde::{Deserialize, Serialize};
-use utoipa::ToSchema;
-
-#[derive(Deserialize)]
-pub struct CreateTokenRequest {
-    pub username: String,
-    pub password: String,
-    #[serde(default = "default_ttl")]
-    pub ttl_days: u64,
-    pub description: Option<String>,
-    #[serde(default = "default_role_str")]
-    pub role: String,
-}
-
-fn default_role_str() -> String {
-    "read".to_string()
-}
-
-fn default_ttl() -> u64 {
-    30
-}
-
-#[derive(Serialize)]
-pub struct CreateTokenResponse {
-    pub token: String,
-    pub expires_in_days: u64,
-}
-
-#[derive(Serialize, ToSchema)]
-#[schema(as = TokenInfo)]
-pub struct TokenListItem {
-    pub hash_prefix: String,
-    pub created_at: u64,
-    pub expires_at: u64,
-    pub last_used: Option<u64>,
-    pub description: Option<String>,
-    pub role: String,
-}
-
-#[derive(Serialize, ToSchema)]
-pub struct TokenListResponse {
-    pub tokens: Vec<TokenListItem>,
-}
-
-/// Create a new API token (requires Basic auth)
-async fn create_token(
-    State(state): State<Arc<AppState>>,
-    ConnectInfo(addr): ConnectInfo<SocketAddr>,
-    headers: HeaderMap,
-    Json(req): Json<CreateTokenRequest>,
-) -> Response {
-    let client_ip = resolve_client_ip(addr.ip(), &headers, &state.config.auth.trusted_proxies);
-    if let Some(retry_after) = state.auth_failures.check_blocked(&client_ip) {
-        return (
-            StatusCode::TOO_MANY_REQUESTS,
-            [(header::RETRY_AFTER, retry_after.to_string())],
-            format!(
-                r#"{{"error":"Too many failed attempts. Retry after {} seconds."}}"#,
-                retry_after
-            ),
-        )
-            .into_response();
-    }
-
-    // Verify user credentials first
-    let auth = match &state.auth {
-        Some(auth) => auth,
-        None => return (StatusCode::SERVICE_UNAVAILABLE, "Auth not configured").into_response(),
-    };
-
-    if !auth.authenticate(&req.username, &req.password) {
-        state.auth_failures.record_failure(client_ip);
-        return (StatusCode::UNAUTHORIZED, "Invalid credentials").into_response();
-    }
-
-    state.auth_failures.record_success(&client_ip);
-
-    let token_store = match &state.tokens {
-        Some(ts) => ts,
-        None => {
-            return (
-                StatusCode::SERVICE_UNAVAILABLE,
-                "Token storage not configured",
-            )
-                .into_response()
-        }
-    };
-
-    let role = match req.role.as_str() {
-        "read" => Role::Read,
-        "write" => Role::Write,
-        "admin" => Role::Admin,
-        _ => {
-            return (
-                StatusCode::BAD_REQUEST,
-                "Invalid role. Use: read, write, admin",
-            )
-                .into_response()
-        }
-    };
-    match token_store.create_token(&req.username, req.ttl_days, req.description, role) {
-        Ok(token) => Json(CreateTokenResponse {
-            token,
-            expires_in_days: req.ttl_days,
-        })
-        .into_response(),
-        Err(e) => (StatusCode::INTERNAL_SERVER_ERROR, e.to_string()).into_response(),
-    }
-}
-
-/// List tokens for authenticated user
-async fn list_tokens(
-    State(state): State<Arc<AppState>>,
-    ConnectInfo(addr): ConnectInfo<SocketAddr>,
-    headers: HeaderMap,
-    Json(req): Json<CreateTokenRequest>,
-) -> Response {
-    let client_ip = resolve_client_ip(addr.ip(), &headers, &state.config.auth.trusted_proxies);
-    if let Some(retry_after) = state.auth_failures.check_blocked(&client_ip) {
-        return (
-            StatusCode::TOO_MANY_REQUESTS,
-            [(header::RETRY_AFTER, retry_after.to_string())],
-            format!(
-                r#"{{"error":"Too many failed attempts. Retry after {} seconds."}}"#,
-                retry_after
-            ),
-        )
-            .into_response();
-    }
-
-    let auth = match &state.auth {
-        Some(auth) => auth,
-        None => return (StatusCode::SERVICE_UNAVAILABLE, "Auth not configured").into_response(),
-    };
-
-    if !auth.authenticate(&req.username, &req.password) {
-        state.auth_failures.record_failure(client_ip);
-        return (StatusCode::UNAUTHORIZED, "Invalid credentials").into_response();
-    }
-
-    state.auth_failures.record_success(&client_ip);
-
-    let token_store = match &state.tokens {
-        Some(ts) => ts,
-        None => {
-            return (
-                StatusCode::SERVICE_UNAVAILABLE,
-                "Token storage not configured",
-            )
-                .into_response()
-        }
-    };
-
-    let tokens: Vec<TokenListItem> = token_store
-        .list_tokens(&req.username)
-        .into_iter()
-        .map(|t| TokenListItem {
-            hash_prefix: t.file_id,
-            created_at: t.created_at,
-            expires_at: t.expires_at,
-            last_used: t.last_used,
-            description: t.description,
-            role: t.role.to_string(),
-        })
-        .collect();
-
-    Json(TokenListResponse { tokens }).into_response()
-}
-
-#[derive(Deserialize)]
-pub struct RevokeRequest {
-    pub username: String,
-    pub password: String,
-    pub hash_prefix: String,
-}
-
-/// Revoke a token
-async fn revoke_token(
-    State(state): State<Arc<AppState>>,
-    ConnectInfo(addr): ConnectInfo<SocketAddr>,
-    headers: HeaderMap,
-    Json(req): Json<RevokeRequest>,
-) -> Response {
-    let client_ip = resolve_client_ip(addr.ip(), &headers, &state.config.auth.trusted_proxies);
-    if let Some(retry_after) = state.auth_failures.check_blocked(&client_ip) {
-        return (
-            StatusCode::TOO_MANY_REQUESTS,
-            [(header::RETRY_AFTER, retry_after.to_string())],
-            format!(
-                r#"{{"error":"Too many failed attempts. Retry after {} seconds."}}"#,
-                retry_after
-            ),
-        )
-            .into_response();
-    }
-
-    let auth = match &state.auth {
-        Some(auth) => auth,
-        None => return (StatusCode::SERVICE_UNAVAILABLE, "Auth not configured").into_response(),
-    };
-
-    if !auth.authenticate(&req.username, &req.password) {
-        state.auth_failures.record_failure(client_ip);
-        return (StatusCode::UNAUTHORIZED, "Invalid credentials").into_response();
-    }
-
-    state.auth_failures.record_success(&client_ip);
-
-    let token_store = match &state.tokens {
-        Some(ts) => ts,
-        None => {
-            return (
-                StatusCode::SERVICE_UNAVAILABLE,
-                "Token storage not configured",
-            )
-                .into_response()
-        }
-    };
-
-    match token_store.revoke_token(&req.hash_prefix) {
-        Ok(()) => (StatusCode::OK, "Token revoked").into_response(),
-        Err(e) => (StatusCode::NOT_FOUND, e.to_string()).into_response(),
-    }
-}
-
-/// Token management routes
-pub fn token_routes() -> Router<Arc<AppState>> {
-    Router::new()
-        .route("/api/tokens", post(create_token))
-        .route("/api/tokens/list", post(list_tokens))
-        .route("/api/tokens/revoke", post(revoke_token))
-}
-
 #[cfg(test)]
 #[allow(clippy::unwrap_used)]
 mod tests {
     use super::*;
-    use std::io::Write;
-    use tempfile::NamedTempFile;
-
-    fn create_test_htpasswd(entries: &[(&str, &str)]) -> NamedTempFile {
-        let mut file = NamedTempFile::new().unwrap();
-        for (username, password) in entries {
-            let hash = bcrypt::hash(password, 4).unwrap(); // cost=4 for speed in tests
-            writeln!(file, "{}:{}", username, hash).unwrap();
-        }
-        file.flush().unwrap();
-        file
-    }
-
-    #[test]
-    fn test_htpasswd_loading() {
-        let file = create_test_htpasswd(&[("admin", "secret"), ("user", "password")]);
-
-        let auth = HtpasswdAuth::from_file(file.path()).unwrap();
-        let users = auth.list_users();
-        assert_eq!(users.len(), 2);
-        assert!(users.contains(&"admin"));
-        assert!(users.contains(&"user"));
-    }
-
-    #[test]
-    fn test_htpasswd_loading_empty_file() {
-        let file = NamedTempFile::new().unwrap();
-        let auth = HtpasswdAuth::from_file(file.path());
-        assert!(auth.is_none());
-    }
-
-    #[test]
-    fn test_htpasswd_loading_with_comments() {
-        let mut file = NamedTempFile::new().unwrap();
-        writeln!(file, "# This is a comment").unwrap();
-        writeln!(file).unwrap();
-        let hash = bcrypt::hash("secret", 4).unwrap();
-        writeln!(file, "admin:{}", hash).unwrap();
-        file.flush().unwrap();
-
-        let auth = HtpasswdAuth::from_file(file.path()).unwrap();
-        assert_eq!(auth.list_users().len(), 1);
-    }
-
-    #[test]
-    fn test_authenticate_valid() {
-        let file = create_test_htpasswd(&[("test", "secret")]);
-        let auth = HtpasswdAuth::from_file(file.path()).unwrap();
-
-        assert!(auth.authenticate("test", "secret"));
-    }
-
-    #[test]
-    fn test_authenticate_invalid_password() {
-        let file = create_test_htpasswd(&[("test", "secret")]);
-        let auth = HtpasswdAuth::from_file(file.path()).unwrap();
-
-        assert!(!auth.authenticate("test", "wrong"));
-    }
-
-    #[test]
-    fn test_authenticate_unknown_user() {
-        let file = create_test_htpasswd(&[("test", "secret")]);
-        let auth = HtpasswdAuth::from_file(file.path()).unwrap();
-
-        assert!(!auth.authenticate("unknown", "secret"));
-    }
 
     #[test]
     fn test_is_public_path() {
@@ -762,47 +424,6 @@ mod tests {
         assert!(!is_public_path("/admin"));
         assert!(!is_public_path("/secret"));
         assert!(!is_public_path("/api/data"));
-    }
-
-    #[test]
-    fn test_default_role_str() {
-        assert_eq!(default_role_str(), "read");
-    }
-
-    #[test]
-    fn test_default_ttl() {
-        assert_eq!(default_ttl(), 30);
-    }
-
-    #[test]
-    fn test_create_token_request_defaults() {
-        let json = r#"{"username":"admin","password":"pass"}"#;
-        let req: CreateTokenRequest = serde_json::from_str(json).unwrap();
-        assert_eq!(req.username, "admin");
-        assert_eq!(req.password, "pass");
-        assert_eq!(req.ttl_days, 30);
-        assert_eq!(req.role, "read");
-        assert!(req.description.is_none());
-    }
-
-    #[test]
-    fn test_create_token_request_custom() {
-        let json = r#"{"username":"admin","password":"pass","ttl_days":90,"role":"write","description":"CI token"}"#;
-        let req: CreateTokenRequest = serde_json::from_str(json).unwrap();
-        assert_eq!(req.ttl_days, 90);
-        assert_eq!(req.role, "write");
-        assert_eq!(req.description, Some("CI token".to_string()));
-    }
-
-    #[test]
-    fn test_create_token_response_serialization() {
-        let resp = CreateTokenResponse {
-            token: "nora_abc123".to_string(),
-            expires_in_days: 30,
-        };
-        let json = serde_json::to_string(&resp).unwrap();
-        assert!(json.contains("nora_abc123"));
-        assert!(json.contains("30"));
     }
 
     #[test]
